+++ Assets/MeshUtility/Runtime/MeshIntegrationResult.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using UnityEngine;
+
+namespace MeshUtility
+{
+    [System.Serializable]
+    public class MeshIntegrationResult
+    {
+        public List<SkinnedMeshRenderer> SourceSkinnedMeshRenderers = new List<SkinnedMeshRenderer>();
+        public List<MeshRenderer> SourceMeshRenderers = new List<MeshRenderer>();
+        public SkinnedMeshRenderer IntegratedRenderer;
+    }
+}
