+++ Assets/MeshUtility/Runtime/MeshIntegrationResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1c66a21d479b3e4a92eedd622d27f4f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
